+++ vivarium/compartments/transport_metabolism.py
@@ -3,8 +3,6 @@ from __future__ import absolute_import, division, print_function
 import os
 import argparse
 
-import matplotlib.pyplot as plt
-
 from vivarium.library.dict_utils import get_value_from_path
 from vivarium.library.units import units
 from vivarium.core.experiment import Compartment
@@ -19,6 +17,7 @@ from vivarium.parameters.parameters import (
     plot_scan_results)
 
 # processes
+from vivarium.plots.transport_metabolism import plot_diauxic_shift
 from vivarium.processes.meta_division import MetaDivision
 from vivarium.processes.metabolism import (
     Metabolism,
@@ -161,7 +160,14 @@ class TransportMetabolism(Compartment):
 
 
 # simulate
-default_test_setting = {
+default_compartment_config = {
+    'external_path': ('external',),
+    'exchange_path': ('exchange',),
+    'global_path': ('global',),
+    'agents_path': ('agents',)}
+
+def test_txp_mtb_ge():
+    default_test_setting = {
         'environment': {
             'volume': 1e-12 * units.L,
             'ports': {
@@ -171,36 +177,34 @@ default_test_setting = {
         'timestep': 1,
         'total_time': 10}
 
-def test_txp_mtb_ge(settings=default_test_setting):
+    compartment = TransportMetabolism(default_compartment_config)
+    return simulate_compartment_in_experiment(compartment, default_test_setting)
 
-    # configure the compartment
-    compartment_config = {
-        'external_path': ('external',),
-        'exchange_path': ('exchange',),
-        'global_path': ('global',),
-        'agents_path': ('agents',)}
-    compartment = TransportMetabolism(compartment_config)
+def simulate_txp_mtb_ge(config=default_compartment_config, out_dir='out'):
 
-    # simulate
-    return simulate_compartment_in_experiment(compartment, settings)
-
-def simulate_txp_mtb_ge(config={}, out_dir='out'):
-
-    end_time = 200  # 2520 sec (42 min) is the expected doubling time in minimal media
+    end_time = 400  # 2520 sec (42 min) is the expected doubling time in minimal media
     timeline = [
         (0, {
             ('external', 'glc__D_e'): 3.0,
             ('external', 'lcts_e'): 3.0,
         }),
         (end_time, {})]
-    glc_lct_settings = {
+
+    sim_settings = {
+        'environment': {
+            'volume': 1e-14 * units.L,
+            'ports': {
+                'external': ('boundary', 'external'),
+                'exchange': ('boundary', 'exchange'),
+            }},
         'timeline': {
             'timeline': timeline,
             'ports': {
                 'external': ('boundary', 'external')}}}
 
     # run simulation
-    timeseries = test_txp_mtb_ge(glc_lct_settings)
+    compartment = TransportMetabolism(default_compartment_config)
+    timeseries = simulate_compartment_in_experiment(compartment, sim_settings)
 
     # calculate growth
     volume_ts = timeseries['boundary']['volume']
@@ -228,86 +232,6 @@ def simulate_txp_mtb_ge(config={}, out_dir='out'):
         'skip_ports': ['null', 'reactions'],
     }
     plot_simulation_output(timeseries, plot_settings, out_dir)
-
-# plots
-def plot_diauxic_shift(timeseries, settings={}, out_dir='out'):
-    external_path = settings.get('external_path', ('environment',))
-    internal_path = settings.get('internal_path', ('cytoplasm',))
-    internal_counts_path = settings.get('internal_counts_path', ('cytoplasm_counts',))
-    reactions_path = settings.get('reactions_path', ('reactions',))
-    global_path = settings.get('global_path', ('global',))
-
-    time = [t/60 for t in timeseries['time']]  # convert to minutes
-
-    environment = get_value_from_path(timeseries, external_path)
-    cell = get_value_from_path(timeseries, internal_path)
-    cell_counts = get_value_from_path(timeseries, internal_counts_path)
-    reactions = get_value_from_path(timeseries, reactions_path)
-    globals = get_value_from_path(timeseries, global_path)
-
-    # environment
-    lactose = environment['lcts_e']
-    glucose = environment['glc__D_e']
-
-    # internal
-    LacY = cell['LacY']
-    lacy_RNA = cell['lacy_RNA']
-    LacY_counts = cell_counts['LacY']
-    lacy_RNA_counts = cell_counts['lacy_RNA']
-
-    # reactions
-    glc_exchange = reactions['EX_glc__D_e']
-    lac_exchange = reactions['EX_lcts_e']
-
-    # global
-    mass = globals['mass']
-
-    # settings
-    environment_volume = settings.get('environment_volume')
-
-    n_cols = 2
-    n_rows = 4
-
-    # make figure and plot
-    fig = plt.figure(figsize=(n_cols * 6, n_rows * 1.5))
-    grid = plt.GridSpec(n_rows, n_cols)
-
-    ax1 = fig.add_subplot(grid[0, 0])  # grid is (row, column)
-    ax1.plot(time, glucose, label='glucose')
-    ax1.plot(time, lactose, label='lactose')
-    set_axes(ax1)
-    ax1.title.set_text('environment, volume = {} L'.format(environment_volume))
-    ax1.set_ylabel('(mM)')
-    ax1.legend(loc='center left', bbox_to_anchor=(1.0, 0.5))
-
-    ax2 = fig.add_subplot(grid[1, 0])  # grid is (row, column)
-    ax2.plot(time, lacy_RNA, label='lacy_RNA')
-    ax2.plot(time, LacY, label='LacY')
-    set_axes(ax2)
-    ax2.title.set_text('internal')
-    ax2.set_ylabel('(mM)')
-    ax2.legend(loc='center left', bbox_to_anchor=(1.0, 0.5))
-
-    ax3 = fig.add_subplot(grid[2, 0])  # grid is (row, column)
-    ax3.plot(time, mass, label='mass')
-    set_axes(ax3, True)
-    ax3.title.set_text('global')
-    ax3.set_ylabel('(fg)')
-    ax3.set_xlabel('time (min)')
-    ax3.legend(loc='center left', bbox_to_anchor=(1.0, 0.5))
-
-    ax4 = fig.add_subplot(grid[0, 1])  # grid is (row, column)
-    ax4.plot(time, glc_exchange, label='glucose exchange')
-    ax4.plot(time, lac_exchange, label='lactose exchange')
-    set_axes(ax4, True)
-    ax4.title.set_text('flux'.format(environment_volume))
-    ax4.set_xlabel('time (min)')
-    ax4.legend(loc='center left', bbox_to_anchor=(1.0, 0.5))
-
-    # save figure
-    fig_path = os.path.join(out_dir, 'diauxic_shift')
-    plt.subplots_adjust(wspace=0.6, hspace=0.5)
-    plt.savefig(fig_path, bbox_inches='tight')
 
 # parameters
 def scan_txp_mtb_ge():
